api/requirements.txt
Django==1.6.1
py3k-bcrypt==0.3
redis==2.8.0
